[ZPI-29] - Change folder structure, correct protos
