+++ backend/clap/src/main/java/zpi/ppea/clap/dtos/AssessorDto.java
@@ -1,19 +1,16 @@
-package zpi.ppea.clap.mapstruct.dtos;
+package zpi.ppea.clap.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AssessorRatingsDto {
-    private int assessorId;
+public class AssessorDto {
     private String firstName;
     private String lastName;
-    private List<PartialRatingDto> partialRatings;
+    private int assessorId;
 }
