+++ backend/clap/src/main/java/zpi/ppea/clap/dtos/AssessorRatingsDto.java
@@ -1,17 +1,19 @@
-package zpi.ppea.clap.mapstruct.dtos;
+package zpi.ppea.clap.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RatingDto {
-    private int ratingId;
-    private boolean isDraft;
+public class AssessorRatingsDto {
     private int assessorId;
-    private RatingType type;
+    private String firstName;
+    private String lastName;
+    private List<PartialRatingDto> partialRatings;
 }
