+++ backend/clap/src/main/java/zpi/ppea/clap/dtos/DetailsSubmissionRequest.java
@@ -1,4 +1,4 @@
-package zpi.ppea.clap.mapstruct.dtos;
+package zpi.ppea.clap.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AssessorDto {
-    private String firstName;
-    private String lastName;
-    private int assessorId;
+public class DetailsSubmissionRequest {
+    private int submissionId;
 }
