+++ backend/clap/src/main/java/zpi/ppea/clap/dtos/DetailsSubmissionResponse.java
@@ -1,4 +1,4 @@
-package zpi.ppea.clap.mapstruct.dtos;
+package zpi.ppea.clap.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
