+++ backend/clap/src/main/java/zpi/ppea/clap/dtos/RatingsSubmissionRequest.java
@@ -1,16 +1,14 @@
-package zpi.ppea.clap.mapstruct.dtos;
+package zpi.ppea.clap.dtos;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class SubmissionsResponse {
-    private List<SubmissionDto> submissionDTOS;
+public class RatingsSubmissionRequest {
+    private int submissionId;
 }
