+++ backend/clap/src/main/java/zpi/ppea/clap/mappers/AssessorDtoMapper.java
@@ -1,8 +1,8 @@
-package zpi.ppea.clap.mapstruct.mappers;
+package zpi.ppea.clap.mappers;
 
 import data_store.Assessor;
 import org.mapstruct.Mapper;
-import zpi.ppea.clap.mapstruct.dtos.AssessorDto;
+import zpi.ppea.clap.dtos.AssessorDto;
 
 @Mapper(componentModel = "spring")
 public interface AssessorDtoMapper {
