+++ frontend/package.json
@@ -33,7 +33,7 @@
     "react-dom": "18.2.0",
     "react-i18next": "^13.3.0",
     "react-icons": "^4.11.0",
-    "sass": "^1.69.4",
+    "sass": "^1.69.5",
     "tailwindcss": "3.3.3",
     "typescript": "5.2.2"
   },
